+++ Piggyback.cs
@@ -1,7 +1,10 @@
-﻿using BepInEx;
+﻿using System;
+using System.Reflection;
+using BepInEx;
+using BepInEx.Configuration;
 using BepInEx.Logging;
 using HarmonyLib;
-using UnityEngine;
+using Object = UnityEngine.Object;
 
 namespace Piggyback;
 
@@ -9,13 +12,51 @@ namespace Piggyback;
 public class Piggyback : BaseUnityPlugin
 {
     internal static new ManualLogSource Logger;
-    private readonly Harmony harmony = new(MyPluginInfo.PLUGIN_GUID);
+
+    private static ConfigEntry<float> s_holdToCarrySetting;
+    private static bool s_isHoldToCarryPatchApplied = false;
+
+    private readonly Harmony m_harmony = new(MyPluginInfo.PLUGIN_GUID);
 
     private void Awake()
     {
         Logger = base.Logger;
-        harmony.PatchAll(typeof(CanBeCarriedPatch));
+
+        SetupConfig();
+
+        m_harmony.PatchAll(typeof(CanBeCarriedPatch));
+        if (s_holdToCarrySetting.Value > 0.0f)
+        {
+            m_harmony.PatchAll(typeof(HoldToCarryPatch));
+            s_isHoldToCarryPatchApplied = true;
+        }
+
         Logger.LogInfo($"Plugin {MyPluginInfo.PLUGIN_NAME} loaded");
+    }
+
+    private void SetupConfig()
+    {
+        s_holdToCarrySetting = Config.Bind("General", "HoldToCarryTime", 1.5f,
+            new ConfigDescription(
+                "The time in seconds you need to hold the interact button to start carrying another player. " +
+                "If 0, you will start carrying the player immediately upon pressing the interact button. " +
+                "If the player is passed out, you will start carrying them immediately even if this is set to a " +
+                "value greater than 0.", new AcceptableValueRange<float>(0.0f, 5.0f)));
+
+        s_holdToCarrySetting.SettingChanged += (sender, args) =>
+        {
+            if (s_holdToCarrySetting.Value <= 0.0f) return;
+            if (s_isHoldToCarryPatchApplied)
+            {
+                Logger.LogInfo($"Hold to Carry time set to {s_holdToCarrySetting.Value} seconds.");
+            }
+            else
+            {
+                m_harmony.PatchAll(typeof(HoldToCarryPatch));
+                s_isHoldToCarryPatchApplied = true;
+                Logger.LogInfo($"Hold to Carry enabled, time set to {s_holdToCarrySetting.Value} seconds.");
+            }
+        };
     }
 
     class CanBeCarriedPatch
@@ -30,17 +71,24 @@ public class Piggyback : BaseUnityPlugin
             // If this is the local character (the player's character itself), don't allow carrying
             if (___character.IsLocal) return;
 
+            // If the character is dead, don't allow carrying
+            if (___character.data.dead) return;
+
             // If the local player has a backpack, don't allow carrying
             if (!Player.localPlayer.backpackSlot.IsEmpty()) return;
+            // if (!Player.localPlayer.backpackSlot.IsEmpty() && !___character.player.backpackSlot.IsEmpty()) return;
+
+            // If the character is carrying another player, don't allow carrying
+            if ((bool)(Object)___character.data.carriedPlayer) return;
+
+            // If the character is already being carried, don't allow carrying
+            if ((bool)(Object)___character.data.carrier) return;
 
             // If the local character is holding an item that can be used on friends, don't allow carrying
             if ((bool)(Object)Character.localCharacter.data.currentItem &&
                 Character.localCharacter.data.currentItem.canUseOnFriend) return;
 
-            if (!___character.data.dead && !(bool)(Object)___character.data.carrier)
-            {
-                __result = true;
-            }
+            __result = true;
         }
 
         [HarmonyPatch(typeof(CharacterCarrying), "Update")]
@@ -56,5 +104,75 @@ public class Piggyback : BaseUnityPlugin
 
             return true;
         }
+    }
+
+    class HoldToCarryPatch
+    {
+        private static readonly Func<CharacterInteractible, bool> CarriedByLocalCharacterDelegate =
+            (Func<CharacterInteractible, bool>)Delegate.CreateDelegate(
+                typeof(Func<CharacterInteractible, bool>),
+                null,
+                typeof(CharacterInteractible).GetMethod("CarriedByLocalCharacter", BindingFlags.Instance | BindingFlags.NonPublic)!
+            );
+
+        [HarmonyPatch(typeof(CharacterInteractible), "Start")]
+        [HarmonyPostfix]
+        static void ReplaceCharacterInteractibleComponent(CharacterInteractible __instance)
+        {
+            var go = __instance.gameObject;
+            if (go.GetComponent<CharacterInteractibleConstant>() != null) return;
+            var newComp = go.AddComponent<CharacterInteractibleConstant>();
+            newComp.character = __instance.character;
+            Destroy(__instance);
+        }
+
+        [HarmonyPatch(typeof(CharacterInteractible), "Interact")]
+        [HarmonyPrefix]
+        static bool CharacterInteractibleInteractPrefix(CharacterInteractible __instance, Character interactor)
+        {
+            return CarriedByLocalCharacterDelegate(__instance)
+                   || __instance.character.data.fullyPassedOut
+                   || s_holdToCarrySetting.Value == 0.0f;
+        }
+    }
+
+    class CharacterInteractibleConstant : CharacterInteractible, IInteractibleConstant
+    {
+        private static readonly Func<CharacterInteractible, bool> CanBeCarriedDelegate =
+            (Func<CharacterInteractible, bool>)Delegate.CreateDelegate(
+                typeof(Func<CharacterInteractible, bool>),
+                null,
+                typeof(CharacterInteractible).GetMethod("CanBeCarried", BindingFlags.Instance | BindingFlags.NonPublic)!
+            );
+        private static readonly Action<CharacterCarrying, Character> StartCarryDelegate =
+            (Action<CharacterCarrying, Character>)Delegate.CreateDelegate(
+                typeof(Action<CharacterCarrying, Character>),
+                null,
+                typeof(CharacterCarrying).GetMethod("StartCarry", BindingFlags.Instance | BindingFlags.NonPublic)!
+            );
+
+        private void Start()
+        {
+            Logger.LogInfo($"CharacterInteractibleConstant added to {character}");
+        }
+
+        public bool IsConstantlyInteractable(Character interactor)
+        {
+            return CanBeCarriedDelegate(this);
+        }
+
+        public float GetInteractTime(Character interactor) => s_holdToCarrySetting.Value;
+
+        public void Interact_CastFinished(Character interactor)
+        {
+            if (CanBeCarriedDelegate(this))
+                StartCarryDelegate(interactor.refs.carriying, character);
+        }
+
+        public void CancelCast(Character interactor) {}
+
+        public void ReleaseInteract(Character interactor) {}
+
+        public bool holdOnFinish => false;
     }
 }