+++ Thunderstore/CHANGELOG.md
@@ -1,3 +1,8 @@
+# v1.1.0
+- Added a config with the following setting:
+  - **HoldToCarryTime**: The time in seconds you need to hold the interact button to start carrying another player. If 0, you will start carrying the player immediately upon pressing the interact button. If the player is passed out, you will start carrying them immediately even if this is set to a value greater than 0. Defaults to 1.5 seconds.
+- Fixed a bug that allowed players to carry each other simultaneously.
+
 # v1.0.1
 - Manifest and Readme changes. No actual changes to the mod itself.